car service

+++ src/main/java/eu/udemx/repository/CarRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarRepository extends JpaRepository<CarDomain, Long> { }
+public interface CarRepository extends JpaRepository<CarDomain, Integer> { }
